+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## 0.9.7
 
 ### Added
+- `GridLine.GetCircleTransform()`
 - `Network.ToModelText(List<Vector3> nodeLocations, Color color)`
-
 - Content Elements can now use an optional disk cache when running locally for testing purposes, to speed up repeated tests or runs, by setting `GltfExtensions.GltfCachePath`.
 
 ### Changed
+- Support non-linear gridlines by deprecating `GridLine.Line` and replacing it with `GridLine.Curve`.
 
 ### Fixed